treat discrete traits (symbols) as integers
consistent with NeXML datatype for symbols.
also updates tests.

+++ R/character_classes.R
@@ -154,13 +154,13 @@ setAs("XMLInternalElementNode", "states",
 ####################################################### 
 ## technically symbol is positive integer http://nexml.org/doc/schema-1/characters/standard/#StandardToken
 setClass("state",
-         slots = c(symbol = "numeric"), 
+         slots = c(symbol = "integer"), 
          contains = "IDTagged")
 setMethod("fromNeXML", 
           signature("state", "XMLInternalElementNode"),
           function(obj, from){
             obj <- callNextMethod()
-            obj@symbol <- as.numeric(xmlAttrs(from)["symbol"])
+            obj@symbol <- as.integer(xmlAttrs(from)["symbol"])
             obj
           })
 setMethod("toNeXML",
